+++ packages/block-logs-stream/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @latticexyz/block-logs-stream
 
+## 2.0.4
+
+### Patch Changes
+
+- Updated dependencies [620e4ec1]
+  - @latticexyz/common@2.0.4
+
+## 2.0.3
+
+### Patch Changes
+
+- Updated dependencies [d2e4d0fb]
+  - @latticexyz/common@2.0.3
+
+## 2.0.2
+
+### Patch Changes
+
+- @latticexyz/common@2.0.2
+
 ## 2.0.1
 
 ### Patch Changes